Removing and cleaning up code
