+++ photocollage/settings/PrintSettings.py
@@ -4,42 +4,6 @@ from reportlab.pdfbase import pdfmetrics
 from reportlab.pdfbase.ttfonts import TTFont
 
 IMAGE_WITH_BLEED_SIZE = (2625, 3375)
-US_LETTER_HARDCOVER = (5700, 3825)
-BOOK_COVER_SIZE = (2588, 3375)
-BOOK_TRIM_SIZE = (2550, 3300)
-COVER_CHILD_IMAGE = (1800, 980)
-
-PPI = 300
-WRAP_AREA_WIDTH_INCH = 0.75
-SAFETY_MARGIN_WIDTH_INCH = 0.5
-BOOK_COVER_SIZE_WIDTH = 8.625
-BOOK_COVER_SIZE_HEIGHT = 11.25
-BOOK_TRIM_SIZE_WIDTH = 8.5
-BOOK_TRIM_SIZE_HEIGHT = 11
-
-SPINE_AREA_WIDTH = 0.25
-SPINE_AREA_WIDTH_PIXEL = SPINE_AREA_WIDTH * PPI
-
-BOOK_COVER_SIZE_WIDTH_PIXEL = BOOK_COVER_SIZE_WIDTH * PPI
-BOOK_COVER_SIZE_HEIGHT_PIXEL = BOOK_COVER_SIZE_HEIGHT * PPI
-BOOK_COVER_SIZE = (BOOK_COVER_SIZE_WIDTH_PIXEL, BOOK_COVER_SIZE_HEIGHT_PIXEL)
-
-BOOK_TRIM_SIZE_WIDTH_PIXEL = BOOK_TRIM_SIZE_WIDTH * PPI
-BOOK_TRIM_SIZE_HEIGHT_PIXEL = BOOK_TRIM_SIZE_HEIGHT * PPI
-
-SAFETY_MARGIN_WIDTH_PIXEL = SAFETY_MARGIN_WIDTH_INCH * PPI
-WRAP_AREA_WIDTH_PIXEL = WRAP_AREA_WIDTH_INCH * PPI
-
-BACK_COVER_TOP_LEFT = (WRAP_AREA_WIDTH_PIXEL, WRAP_AREA_WIDTH_PIXEL)  # (225,225)
-
-BACK_COVER_BOTTOM_RIGHT = (BACK_COVER_TOP_LEFT[0] + BOOK_COVER_SIZE_WIDTH_PIXEL,
-                           US_LETTER_HARDCOVER[1] - WRAP_AREA_WIDTH_PIXEL)
-
-FRONT_COVER_TOP_LEFT = (BACK_COVER_BOTTOM_RIGHT[0] + SPINE_AREA_WIDTH_PIXEL,
-                        BACK_COVER_TOP_LEFT[1])
-
-FRONT_COVER_BOTTOM_RIGHT = (US_LETTER_HARDCOVER[0] - WRAP_AREA_WIDTH_PIXEL,
-                            US_LETTER_HARDCOVER[1] - WRAP_AREA_WIDTH_PIXEL )
 
 FONT_FOLDER = os.path.join('/Users', getpass.getuser(), 'GoogleDrive', 'Fonts')
